Add files via upload
Team Members- Aman Rana, Amit Kumar , Anjali Navrang.
Aman Rana-Working on user panel as well as main page.
Amit Kumar- Working on Database,Feedback analysis and main page.
Anjali Navrang- working on Admin panel as well as main page.

The project is based on E-commerce product feedback analysis. So we have created the website for an e-commerce work as we uploaded the code .now i am uploading my module which is based on user work so that is the part of the website also it can give the user to access his/her account and check the product which he/she seen before and what he/she wants to do with this like buying or not .The main role is in the project is only showing the user feedback analysis how the user give the feedback and analysis it before buying any product.